Initial Problem Thoughts
This is the first run through of me thinking of the problem. This was sent in an email to a member of timescale's frontend team as well. This document only shows my initial thinking and some of the things that came to mind when I was first reading the problem.

Signed-off-by: Dimitri Hector <73374296+joshuahector86@users.noreply.github.com>